taking event function to anouther common file and exporting it to web.mjs
